Consolidating legal app into home app
