+++ fec/home/templates/home/legal_resources_landing.html
@@ -20,8 +20,7 @@
             {% endif %}
           </select>
           <input type="text" name="search" class="combo__input js-search-input"
-          aria-label="Search"
-          {% if search_query %} value="{{ search_query }}"{% endif %}>
+          aria-label="Search">
           <button type="submit" class="button--standard button--search combo__button">
            <span class="u-visually-hidden">Search</span>
           </button>
