+++ fec/home/views.py
@@ -130,3 +130,24 @@ def contact(request):
   return render(request, 'home/contact.html', {
     'self': page_context,
   })
+
+def ao_process(request):
+  ancestors = [
+    {
+      'title': 'Legal resources',
+      'url': '/legal-resources/',
+    }, {
+      'title': 'Advisory opinions',
+      'url': settings.FEC_APP_URL + '/legal/advisory-opinions',
+    }
+  ]
+  page_context = {
+    'content_section': 'legal',
+    'title': 'The advisory opinion process',
+    'ancestors': ancestors
+  }
+
+  return render(request, 'legal/ao_process.html', {
+    'self': page_context
+  })
+
